Feature/backend/order-api/status

+++ OrderApi/Repository/OrderRepository.cs
@@ -1,7 +1,7 @@
-﻿using Microsoft.EntityFrameworkCore;
+﻿using Library.Extensions;
+using Microsoft.EntityFrameworkCore;
 using OrderApi.Data;
 using OrderApi.Models;
-using OrderApi.Models.Extensions;
 
 namespace OrderApi.Repository
 {
@@ -87,7 +87,9 @@ namespace OrderApi.Repository
 
         public async Task<Order?> GetByIdAsync(Guid id)
         {
-            var order = await _context.Orders.FirstOrDefaultAsync(o => o.OrderId == id);
+            var order = await _context.Orders
+                .AsNoTracking()
+                .FirstOrDefaultAsync(o => o.OrderId == id);
             if (order == null)
             {
                 _message = $"Order with Id [{id}] not found.";
@@ -148,21 +150,15 @@ namespace OrderApi.Repository
             _logger.LogInformation($"Order with Id[{order.OrderId}] updated succesfully.");
         }
 
-        public async Task DeleteAsync(Order order)
+        public async Task DeleteAsync(Guid id)
         {
-            try
-            {
-                _context.Orders.Remove(order);
-                await _context.SaveChangesAsync();
-            }
-            catch (Exception ex)
-            {
-                _message = $"Deletion of Order with id [{order.OrderId}] has failed.";
-                _logger.LogError(_message);
-                throw new ArgumentException(_message, ex);
-            }
+            var order = await _context.Orders.FindAsync(id);
 
-            _logger.LogInformation($"Order with Id [{order.OrderId}] deleted succesfully.");
+            if (order == null)
+                throw new KeyNotFoundException();
+
+            _context.Orders.Remove(order);
+            await _context.SaveChangesAsync();
         }
     }
